+++ src/records/TranslationRecord.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace lhs\craftcms\modules\translations\records;
+namespace nstcactus\craftcms\modules\translations\records;
 
 use yii\db\ActiveQuery;
 use craft\db\ActiveRecord;
-use lhs\craftcms\modules\translations\TranslationsModule;
+use nstcactus\craftcms\modules\translations\TranslationsModule;
 
 /**
  * @property int $id